Marine_Anime

+++ DefaultWindow/CMainGame.h
@@ -25,7 +25,7 @@ public:
 private:
 	HDC			m_hDC;
 
-	DWORD		m_dwTime;
+	ULONGLONG	m_dwTime;
 	int			m_iFPS;
 	TCHAR		m_szFPS[16];
